pyproject.toml
[tool.poetry]
name = "api-workshop"
version = "0.1.0"
description = ""
authors = ["Felipe Sebben <felipesebben@yahoo.com.br>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
taskipy = "^1.12.2"
pytest = "^7.4.3"
httpx = "^0.25.2"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
pre-commit = "^3.5.0"
psycopg2 = "^2.9.9"
isort = "^5.13.2"
black = "^23.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]

run = "uvicorn app.main:app --reload"
test = "pytest -v"
kill = "kill -9 $(lsof -t -i:8000)"
format = "isort . && black ."
commit = "git add . && pre-commit run"
